Finished button generation modularization.

+++ include/Utilites.h
@@ -1,6 +1,6 @@
 #ifndef UTILITES_H
 #define UTILITES_H
 
-int parseUTF8(std::vector<std::string> &dst, std::string path);
+size_t parseUTF8(std::vector<std::string> &dst, std::string path);
 
 #endif // UTILITES_H
